requirement.txt
waitress
django-mathfilters
requests
django-braces
django-widget-tweaks